[Modify] grades models

+++ app/db/models/grades.py
@@ -1,6 +1,5 @@
 import datetime
-from decimal import Decimal
-from sqlalchemy import Column, Integer, Numeric, String, DateTime
+from sqlalchemy import Column, Integer, DateTime
 
 from db.base_class import Base
 
@@ -8,5 +7,5 @@ class Grade(Base):
     grade_no = Column(Integer, primary_key=True, autoincrement=True)
     user_no = Column(Integer)
     book_no = Column(Integer)
-    grade = Column(Numeric)
+    grade = Column(Integer)
     createtime = Column(DateTime, default=datetime.datetime.now())
